requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
branca==0.7.2
build==1.2.1
buildozer==1.5.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
docutils==0.21.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.49.0
fpdf==1.7.2
fqdn==1.5.1
fsspec==2024.3.0
functions-framework==3.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-tools==1.62.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
h2o==3.46.0.2
h5py==3.11.0
hpack==4.0.0
htmlmin==0.1.12
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
hyperframe==6.0.1
idna==3.6
ImageHash==4.3.1
imbalanced-learn==0.12.3
import-ipynb==0.1.4
importlib_metadata==7.0.2
importlib_resources==6.3.0
ipykernel==6.29.3
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.3.3
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
language-tool-python==2.7.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.42.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mido==1.3.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multimethod==1.11.2
mysql-connector-python==8.4.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
plyer==2.1.0
pooch==1.8.1
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
PyDrive==1.3.1
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySocks==1.7.1
pyspellchecker==0.8.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.8.0
redis==5.0.4
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.1.post1
scikit-surprise==1.1.4
scipy==1.12.0
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.2
sh==1.14.3
shap==0.45.1
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spotipy==2.23.0
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.35.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.41.1
trio==0.25.1
trio-websocket==0.11.1
typeguard==4.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
visions==0.7.6
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wget==3.2
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.4.0
ydata-profiling==4.8.3
zipp==3.18.0
